[CCUBE-1545][MAHI]Adjust typings and props for link and also remove calculation for paragraph display
